+++ admin/metaboxes/background/metabox-background.php
@@ -78,6 +78,10 @@ function pw_background_meta_save( $post_id ){
     if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;
 
+    // Security Layer 
+    if ( !current_user_can( 'edit_post', $post_id ) )
+        return $post_id;
+
     // Get the posted JSON string    
 	$post = pw_get_posted_json( 'pw_background_post' );