+++ admin/metaboxes/event/metabox-event.php
@@ -131,6 +131,10 @@ function pw_event_meta_save( $post_id ){
 	if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;
 
+    // Security Layer 
+    if ( !current_user_can( 'edit_post', $post_id ) )
+        return $post_id;
+
 	// Get the JSON string which represents the post to be saved 
 	$post = $_POST['pw_event_post'];