+++ admin/metaboxes/layout/metabox-layout.php
@@ -79,6 +79,10 @@ function pw_layout_meta_save( $post_id ){
 	// Stop autosave to preserve meta data
     if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;
+
+    // Security Layer 
+    if ( !current_user_can( 'edit_post', $post_id ) )
+        return $post_id;
     
     // Get the posted JSON string    
     $post = pw_get_posted_json( 'pw_layout_post' );