+++ admin/metaboxes/options/metabox-options.php
@@ -61,6 +61,10 @@ function pw_metabox_options_save( $post_id ){
 	if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;
 
+    // Security Layer 
+    if ( !current_user_can( 'edit_post', $post_id ) )
+        return $post_id;
+
 	$meta_key = PW_POSTMETA_KEY;
 
     $pwMeta = $_POST[ PW_POSTMETA_KEY ];