+++ admin/metaboxes/post_parent/metabox-post_parent.php
@@ -163,6 +163,10 @@ function pw_post_parent_meta_save( $post_id ){
 	if ( wp_is_post_autosave( $post_id ) || wp_is_post_revision( $post_id ) )
         return $post_id;
 
+    // Security Layer 
+    if ( !current_user_can( 'edit_post', $post_id ) )
+        return $post_id;
+
     //pw_log( 'pw_post_parent_meta_save : ' . $post_id );
 
 	// Get the JSON string which represents the post to be saved